# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Project Argus Border Detection System API
  description: |
    Comprehensive API for Project Argus - a world-class border security detection and monitoring solution.
    
    This API provides endpoints for:
    - Real-time detection and tracking management
    - Alert and incident handling
    - Evidence management and forensics
    - System health monitoring
    - User authentication and authorization
    
    ## Authentication
    All endpoints require JWT authentication with role-based access control.
    
    ## Rate Limiting
    API requests are rate-limited to prevent abuse and ensure system stability.
    
  version: 1.0.0
  contact:
    name: Project Argus Team
    email: support@projectargus.com
  license:
    name: Proprietary
servers:
  - url: https://api.projectargus.com/v1
    description: Production server
  - url: https://staging-api.projectargus.com/v1
    description: Staging server
  - url: http://localhost:8000/v1
    description: Development server
security:
  - BearerAuth: []
paths:
...
  /alerts:
    get:
      tags:
        - Alerts
      summary: List alerts
      description: Retrieve alerts with filtering and pagination
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, acknowledged, resolved]
        - name: severity
          in: query
          schema:
            type: string
            enum: [low, medium, high, critical]
        - name: camera_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Alert'
  /alerts/{alert_id}/acknowledge:
    post:
      tags:
        - Alerts
      summary: Acknowledge alert
      description: Mark alert as acknowledged by operator
      parameters:
        - name: alert_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                notes:
                  type: string
                  description: Optional notes from operator
      responses:
        '200':
          description: Alert acknowledged
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Alert'

  # Incident endpoints
  /incidents:
    get:
      tags:
        - Incidents
      summary: List incidents
      description: Retrieve incidents with filtering options
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [open, in_progress, resolved, closed]
        - name: priority
          in: query
          schema:
            type: string
            enum: [low, medium, high, critical]
      responses:
        '200':
          description: List of incidents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Incident'

    post:
      tags:
        - Incidents
      summary: Create incident
      description: Create new incident from alert
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - alert_id
                - title
              properties:
                alert_id:
                  type: string
                  format: uuid
                title:
                  type: string
                description:
                  type: string
                priority:
                  type: string
                  enum: [low, medium, high, critical]
      responses:
        '201':
          description: Incident created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Incident'

  # Evidence endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Alert:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [crossing, loitering, tamper, system_error]
        severity:
          type: string
          enum: [low, medium, high, critical]
        camera_id:
          type: string
          format: uuid
        detection_id:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
        confidence:
          type: number
          minimum: 0
          maximum: 1
        risk_score:
          type: number
          minimum: 0
          maximum: 1
        acknowledged:
          type: boolean
        acknowledged_by:
          type: string
          format: uuid
        acknowledged_at:
          type: string
          format: date-time
        metadata:
          type: object
    Incident:
      type: object
      properties:
        id:
          type: string
          format: uuid
        alert_id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [open, in_progress, resolved, closed]
        priority:
          type: string
          enum: [low, medium, high, critical]
        operator_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        closed_at:
          type: string
          format: date-time
        evidence_ids:
          type: array
          items:
            type: string
            format: uuid
        notes:
          type: array
          items:
            $ref: '#/components/schemas/Note'
...
    Note:
      type: object
      properties:
        id:
          type: string
          format: uuid
        author_id:
          type: string
          format: uuid
        content:
          type: string
        created_at:
          type: string
          format: date-time
...